feat(backend): add client proxy
The proxy consists of a wrapper to the spotify api. It is structured in two main services: Authentication and Tracks service.
The intention is to force the frontend to only make requests to the Tracks service instead of doing it directly to the spotify api, so the spotify api secret key stays secret. The authentication service serves only as a dependency of the Tracks service